+++ src/wifi/model/rrpaa-wifi-manager.cc
@@ -296,7 +296,7 @@ RrpaaWifiManager::InitThresholds (RrpaaWifiRemoteStation *station)
           mtl = nextMtl;
         }
       WifiRrpaaThresholds th;
-      th.m_ewnd = ceil (m_tau / totalTxTime.GetSeconds ());
+      th.m_ewnd = static_cast<uint32_t> (ceil (m_tau / totalTxTime.GetSeconds ()));
       th.m_ori = ori;
       th.m_mtl = mtl;
       station->m_thresholds.push_back (std::make_pair (th, mode));
